% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-jdbc.R
\name{dbConnect,DrillJDBCDriver-method}
\alias{dbConnect,DrillJDBCDriver-method}
\alias{dbConnect}
\title{Connect to a schema/table}
\usage{
\S4method{dbConnect}{DrillJDBCDriver}(drv, url, user = "", password = "", ...)
}
\arguments{
\item{drv}{driver}

\item{url}{connection string}

\item{user, password}{creds}

\item{...}{extra params}
}
\description{
Connect to a schema/table
}
