% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-jdbc.R
\name{dbGetRowsAffected,DrillJDBCResult-method}
\alias{dbGetRowsAffected,DrillJDBCResult-method}
\alias{dbGetRowsAffected}
\title{Drill JDBC dbGetRowsAffected}
\usage{
\S4method{dbGetRowsAffected}{DrillJDBCResult}(res, ...)
}
\arguments{
\item{res}{A \code{\linkS4class{DrillJDBCResult}} object}

\item{...}{Extra optional parameters}
}
\description{
Drill JDBC dbGetRowsAffected
}
