% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-jdbc.R
\name{dbQuoteIdentifier,DrillJDBCConnection,character-method}
\alias{dbQuoteIdentifier,DrillJDBCConnection,character-method}
\alias{dbQuoteIdentifier}
\title{Thin wrapper for dbQuoteIdentifier}
\usage{
\S4method{dbQuoteIdentifier}{DrillJDBCConnection,character}(conn, x, ...)
}
\arguments{
\item{conn}{connection}

\item{x, }{item to quote}

\item{...}{passed on to downstream methods}
}
\description{
Thin wrapper for dbQuoteIdentifier
}
