% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-jdbc.R
\name{dbSendQuery,DrillJDBCConnection,character-method}
\alias{dbSendQuery,DrillJDBCConnection,character-method}
\alias{dbSendQuery}
\title{Thin wrapper for dbSendQuery}
\usage{
\S4method{dbSendQuery}{DrillJDBCConnection,character}(conn, statement, ..., list = NULL)
}
\arguments{
\item{conn}{connection}

\item{statement}{SQL statement,}

\item{...}{passed on}

\item{list}{list}
}
\description{
Thin wrapper for dbSendQuery
}
