% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-jdbc.R
\docType{class}
\name{DrillJDBCDriver-class}
\alias{DrillJDBCDriver-class}
\alias{DrillJDBC}
\alias{DrillJDBCConnection-class}
\alias{DrillJDBCResult-class}
\title{Create a Drill JDBC connection}
\usage{
DrillJDBC(
  identifier.quote = "`",
  drill_jdbc_jar_path = Sys.getenv("DRILL_JDBC_JAR")
)
}
\arguments{
\item{identifier.quote}{quote}

\item{drill_jdbc_jar_path}{path}
}
\description{
Create a Drill JDBC connection
}
\keyword{internal}
