% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant-caffeinated-package.R
\name{sergeant-caffeinated-exports}
\alias{sergeant-caffeinated-exports}
\alias{\%>\%}
\title{sergeant exported operators}
\description{
The following functions are imported and then re-exported
from the sergeant package to enable use of the magrittr
pipe operator with no additional library calls
}
