% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant-caffeinated-package.R
\docType{package}
\name{sergeant-caffeinated}
\alias{sergeant-caffeinated}
\alias{sergeant-caffeinated-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/sergeant-caffeinated}
\item BugReports: \url{https://gitlab.com/hrbrmstr/sergeant-caffeinated/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
