% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-jdbc.R
\name{sql_translate_env.DrillJDBCConnection}
\alias{sql_translate_env.DrillJDBCConnection}
\title{Thin wrapper for sql_translate_env}
\usage{
\method{sql_translate_env}{DrillJDBCConnection}(con)
}
\arguments{
\item{con}{connection}
}
\description{
Thin wrapper for sql_translate_env
}
\keyword{internal}
